fix struct alignment in transposition.go
I wasn't aware that the go complier aligns the structs to met cpu
alignment requirements(at least i learned something new). So it was adding a
padding i was not aware of. Now the transposition table size allocated mets the alignment requirements and
has the correct size. Also for the pawn hash table, added an
explicit(unused) padding to make the struct align to the cpu.